+++ communication/rfid.h
@@ -5,7 +5,7 @@
 #ifndef ARCTOS_ROBOT_RFID_H
 #define ARCTOS_ROBOT_RFID_H
 
-
+#include <unistd.h>
 #include <stddef.h>
 #include "../general.h"
 
@@ -15,6 +15,6 @@
 
 int rfid_init(serial_port_options_t* options);
 
-int rfid_read(int fd, char* buffer, size_t buffer_size);
+ssize_t rfid_read(int fd, char* buffer, size_t buffer_size);
 
 #endif //ARCTOS_ROBOT_RFID_H
